Refactor Cover to reflect dpl-react + Cleanup of Types
