+++ src/stories/Library/material-card/MaterialCard.tsx
@@ -1,7 +1,6 @@
 import { useState } from "react";
 import { ButtonUI } from "../Buttons/button-ui/ButtonUI";
-import Cover from "../cover/Cover";
-import { CoverProps } from "../cover/types";
+import Cover, { CoverProps } from "../cover/Cover";
 
 export type MaterialCardProps = {
   title: string;
